feat: async support for request method
We added an async support for the request
method by creating an async_request method.
It provides the same arguments as the other one.
In order to support async calls, we used the aiohttp module.
We didn't change the way the method works but just changed sync calls
by async calls. It concerns api calls and sleep.